Move file and fix typo.
Still need to fix the CI.
